[MOVE] ImageType, BannerType을 domain 패키지로 이동
