+++ operation-api/src/main/java/org/sopt/makers/operation/web/banner/service/BannerService.java
@@ -1,6 +1,6 @@
 package org.sopt.makers.operation.web.banner.service;
 
-import org.sopt.makers.operation.web.banner.dto.request.*;
+import org.sopt.makers.operation.banner.domain.*;
 import org.sopt.makers.operation.web.banner.dto.response.BannerResponse;
 
 public interface BannerService {
